+++ src/components/Toast/Toast.stories.tsx
@@ -9,7 +9,7 @@ export default { title: 'Components/Toast', component: Toast, parameters: { read
 export const Basic = () => (
   <div style={{ display: 'flex' }}>
     <Toast
-      toastType={select('toastType', ['info', 'success', 'warning', 'error'], 'info')}
+      message={select('message', ['info', 'success', 'warning', 'error'], 'info')}
       heading={text('heading', 'Information about the toast')}
       onClose={boolean('has onClose', false) ? action('onClose') : undefined}
       action={boolean('has action', false) ? { text: 'Action', onClick: action('action.onClick') } : undefined}